Move code to own file
This code does not work. Hostile does not lock the file and only adds
one entry when you have mutriple ones. We should go back to the old code
that I made that was a bit more messy but sync, as async does not work.
